ignore pytest's deprecation warning for fspath in NutsYamlFile
reason: pytest moves from py.path.local to pathlib.Path and will support both for "quite a while" according to pytest:
https://docs.pytest.org/en/latest/deprecations.html#node-fspath-in-favor-of-pathlib-and-node-path

We do not introduce a compatibility layer for both since it would duplicate a lot of code just for the sake of distinction between py.path.local and pathlib.Path. Nuts will only change the code if the legacy py.path.local is also dropped in pytest, so that older pytest versions can be used with nuts too until then.